+++ apps/c/CloverLeaf/OpenACC/PdV_kernel_predict_openacc_kernel.cpp
@@ -107,121 +107,181 @@ void ops_par_loop_PdV_kernel_predict(char const *name, ops_block block, int dim,
 
 
   //set up initial pointers
-  int base0 = args[0].dat->base_offset + (block->instance->OPS_soa ? args[0].dat->type_size : args[0].dat->elem_size) * start[0] * args[0].stencil->stride[0];
-  base0 = base0 + (block->instance->OPS_soa ? args[0].dat->type_size : args[0].dat->elem_size) *
-    args[0].dat->size[0] *
-    start[1] * args[0].stencil->stride[1];
-  #ifdef OPS_GPU
+  long long int base0 =
+      args[0].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[0].dat->type_size
+                                                     : args[0].dat->elem_size) *
+                                     start[0] * args[0].stencil->stride[0];
+  base0 = base0 +
+          (long long int)(block->instance->OPS_soa ? args[0].dat->type_size
+                                                   : args[0].dat->elem_size) *
+              args[0].dat->size[0] * start[1] * args[0].stencil->stride[1];
+#ifdef OPS_GPU
   double *p_a0 = (double *)((char *)args[0].data_d + base0);
   #else
   double *p_a0 = (double *)((char *)args[0].data + base0);
   #endif
 
-  int base1 = args[1].dat->base_offset + (block->instance->OPS_soa ? args[1].dat->type_size : args[1].dat->elem_size) * start[0] * args[1].stencil->stride[0];
-  base1 = base1 + (block->instance->OPS_soa ? args[1].dat->type_size : args[1].dat->elem_size) *
-    args[1].dat->size[0] *
-    start[1] * args[1].stencil->stride[1];
-  #ifdef OPS_GPU
+  long long int base1 =
+      args[1].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[1].dat->type_size
+                                                     : args[1].dat->elem_size) *
+                                     start[0] * args[1].stencil->stride[0];
+  base1 = base1 +
+          (long long int)(block->instance->OPS_soa ? args[1].dat->type_size
+                                                   : args[1].dat->elem_size) *
+              args[1].dat->size[0] * start[1] * args[1].stencil->stride[1];
+#ifdef OPS_GPU
   double *p_a1 = (double *)((char *)args[1].data_d + base1);
   #else
   double *p_a1 = (double *)((char *)args[1].data + base1);
   #endif
 
-  int base2 = args[2].dat->base_offset + (block->instance->OPS_soa ? args[2].dat->type_size : args[2].dat->elem_size) * start[0] * args[2].stencil->stride[0];
-  base2 = base2 + (block->instance->OPS_soa ? args[2].dat->type_size : args[2].dat->elem_size) *
-    args[2].dat->size[0] *
-    start[1] * args[2].stencil->stride[1];
-  #ifdef OPS_GPU
+  long long int base2 =
+      args[2].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[2].dat->type_size
+                                                     : args[2].dat->elem_size) *
+                                     start[0] * args[2].stencil->stride[0];
+  base2 = base2 +
+          (long long int)(block->instance->OPS_soa ? args[2].dat->type_size
+                                                   : args[2].dat->elem_size) *
+              args[2].dat->size[0] * start[1] * args[2].stencil->stride[1];
+#ifdef OPS_GPU
   double *p_a2 = (double *)((char *)args[2].data_d + base2);
   #else
   double *p_a2 = (double *)((char *)args[2].data + base2);
   #endif
 
-  int base3 = args[3].dat->base_offset + (block->instance->OPS_soa ? args[3].dat->type_size : args[3].dat->elem_size) * start[0] * args[3].stencil->stride[0];
-  base3 = base3 + (block->instance->OPS_soa ? args[3].dat->type_size : args[3].dat->elem_size) *
-    args[3].dat->size[0] *
-    start[1] * args[3].stencil->stride[1];
-  #ifdef OPS_GPU
+  long long int base3 =
+      args[3].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[3].dat->type_size
+                                                     : args[3].dat->elem_size) *
+                                     start[0] * args[3].stencil->stride[0];
+  base3 = base3 +
+          (long long int)(block->instance->OPS_soa ? args[3].dat->type_size
+                                                   : args[3].dat->elem_size) *
+              args[3].dat->size[0] * start[1] * args[3].stencil->stride[1];
+#ifdef OPS_GPU
   double *p_a3 = (double *)((char *)args[3].data_d + base3);
   #else
   double *p_a3 = (double *)((char *)args[3].data + base3);
   #endif
 
-  int base4 = args[4].dat->base_offset + (block->instance->OPS_soa ? args[4].dat->type_size : args[4].dat->elem_size) * start[0] * args[4].stencil->stride[0];
-  base4 = base4 + (block->instance->OPS_soa ? args[4].dat->type_size : args[4].dat->elem_size) *
-    args[4].dat->size[0] *
-    start[1] * args[4].stencil->stride[1];
-  #ifdef OPS_GPU
+  long long int base4 =
+      args[4].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[4].dat->type_size
+                                                     : args[4].dat->elem_size) *
+                                     start[0] * args[4].stencil->stride[0];
+  base4 = base4 +
+          (long long int)(block->instance->OPS_soa ? args[4].dat->type_size
+                                                   : args[4].dat->elem_size) *
+              args[4].dat->size[0] * start[1] * args[4].stencil->stride[1];
+#ifdef OPS_GPU
   double *p_a4 = (double *)((char *)args[4].data_d + base4);
   #else
   double *p_a4 = (double *)((char *)args[4].data + base4);
   #endif
 
-  int base5 = args[5].dat->base_offset + (block->instance->OPS_soa ? args[5].dat->type_size : args[5].dat->elem_size) * start[0] * args[5].stencil->stride[0];
-  base5 = base5 + (block->instance->OPS_soa ? args[5].dat->type_size : args[5].dat->elem_size) *
-    args[5].dat->size[0] *
-    start[1] * args[5].stencil->stride[1];
-  #ifdef OPS_GPU
+  long long int base5 =
+      args[5].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[5].dat->type_size
+                                                     : args[5].dat->elem_size) *
+                                     start[0] * args[5].stencil->stride[0];
+  base5 = base5 +
+          (long long int)(block->instance->OPS_soa ? args[5].dat->type_size
+                                                   : args[5].dat->elem_size) *
+              args[5].dat->size[0] * start[1] * args[5].stencil->stride[1];
+#ifdef OPS_GPU
   double *p_a5 = (double *)((char *)args[5].data_d + base5);
   #else
   double *p_a5 = (double *)((char *)args[5].data + base5);
   #endif
 
-  int base6 = args[6].dat->base_offset + (block->instance->OPS_soa ? args[6].dat->type_size : args[6].dat->elem_size) * start[0] * args[6].stencil->stride[0];
-  base6 = base6 + (block->instance->OPS_soa ? args[6].dat->type_size : args[6].dat->elem_size) *
-    args[6].dat->size[0] *
-    start[1] * args[6].stencil->stride[1];
-  #ifdef OPS_GPU
+  long long int base6 =
+      args[6].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[6].dat->type_size
+                                                     : args[6].dat->elem_size) *
+                                     start[0] * args[6].stencil->stride[0];
+  base6 = base6 +
+          (long long int)(block->instance->OPS_soa ? args[6].dat->type_size
+                                                   : args[6].dat->elem_size) *
+              args[6].dat->size[0] * start[1] * args[6].stencil->stride[1];
+#ifdef OPS_GPU
   double *p_a6 = (double *)((char *)args[6].data_d + base6);
   #else
   double *p_a6 = (double *)((char *)args[6].data + base6);
   #endif
 
-  int base7 = args[7].dat->base_offset + (block->instance->OPS_soa ? args[7].dat->type_size : args[7].dat->elem_size) * start[0] * args[7].stencil->stride[0];
-  base7 = base7 + (block->instance->OPS_soa ? args[7].dat->type_size : args[7].dat->elem_size) *
-    args[7].dat->size[0] *
-    start[1] * args[7].stencil->stride[1];
-  #ifdef OPS_GPU
+  long long int base7 =
+      args[7].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[7].dat->type_size
+                                                     : args[7].dat->elem_size) *
+                                     start[0] * args[7].stencil->stride[0];
+  base7 = base7 +
+          (long long int)(block->instance->OPS_soa ? args[7].dat->type_size
+                                                   : args[7].dat->elem_size) *
+              args[7].dat->size[0] * start[1] * args[7].stencil->stride[1];
+#ifdef OPS_GPU
   double *p_a7 = (double *)((char *)args[7].data_d + base7);
   #else
   double *p_a7 = (double *)((char *)args[7].data + base7);
   #endif
 
-  int base8 = args[8].dat->base_offset + (block->instance->OPS_soa ? args[8].dat->type_size : args[8].dat->elem_size) * start[0] * args[8].stencil->stride[0];
-  base8 = base8 + (block->instance->OPS_soa ? args[8].dat->type_size : args[8].dat->elem_size) *
-    args[8].dat->size[0] *
-    start[1] * args[8].stencil->stride[1];
-  #ifdef OPS_GPU
+  long long int base8 =
+      args[8].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[8].dat->type_size
+                                                     : args[8].dat->elem_size) *
+                                     start[0] * args[8].stencil->stride[0];
+  base8 = base8 +
+          (long long int)(block->instance->OPS_soa ? args[8].dat->type_size
+                                                   : args[8].dat->elem_size) *
+              args[8].dat->size[0] * start[1] * args[8].stencil->stride[1];
+#ifdef OPS_GPU
   double *p_a8 = (double *)((char *)args[8].data_d + base8);
   #else
   double *p_a8 = (double *)((char *)args[8].data + base8);
   #endif
 
-  int base9 = args[9].dat->base_offset + (block->instance->OPS_soa ? args[9].dat->type_size : args[9].dat->elem_size) * start[0] * args[9].stencil->stride[0];
-  base9 = base9 + (block->instance->OPS_soa ? args[9].dat->type_size : args[9].dat->elem_size) *
-    args[9].dat->size[0] *
-    start[1] * args[9].stencil->stride[1];
-  #ifdef OPS_GPU
+  long long int base9 =
+      args[9].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[9].dat->type_size
+                                                     : args[9].dat->elem_size) *
+                                     start[0] * args[9].stencil->stride[0];
+  base9 = base9 +
+          (long long int)(block->instance->OPS_soa ? args[9].dat->type_size
+                                                   : args[9].dat->elem_size) *
+              args[9].dat->size[0] * start[1] * args[9].stencil->stride[1];
+#ifdef OPS_GPU
   double *p_a9 = (double *)((char *)args[9].data_d + base9);
   #else
   double *p_a9 = (double *)((char *)args[9].data + base9);
   #endif
 
-  int base10 = args[10].dat->base_offset + (block->instance->OPS_soa ? args[10].dat->type_size : args[10].dat->elem_size) * start[0] * args[10].stencil->stride[0];
-  base10 = base10 + (block->instance->OPS_soa ? args[10].dat->type_size : args[10].dat->elem_size) *
-    args[10].dat->size[0] *
-    start[1] * args[10].stencil->stride[1];
-  #ifdef OPS_GPU
+  long long int base10 =
+      args[10].dat->base_offset +
+      (long long int)(block->instance->OPS_soa ? args[10].dat->type_size
+                                               : args[10].dat->elem_size) *
+          start[0] * args[10].stencil->stride[0];
+  base10 = base10 +
+           (long long int)(block->instance->OPS_soa ? args[10].dat->type_size
+                                                    : args[10].dat->elem_size) *
+               args[10].dat->size[0] * start[1] * args[10].stencil->stride[1];
+#ifdef OPS_GPU
   double *p_a10 = (double *)((char *)args[10].data_d + base10);
   #else
   double *p_a10 = (double *)((char *)args[10].data + base10);
   #endif
 
-  int base11 = args[11].dat->base_offset + (block->instance->OPS_soa ? args[11].dat->type_size : args[11].dat->elem_size) * start[0] * args[11].stencil->stride[0];
-  base11 = base11 + (block->instance->OPS_soa ? args[11].dat->type_size : args[11].dat->elem_size) *
-    args[11].dat->size[0] *
-    start[1] * args[11].stencil->stride[1];
-  #ifdef OPS_GPU
+  long long int base11 =
+      args[11].dat->base_offset +
+      (long long int)(block->instance->OPS_soa ? args[11].dat->type_size
+                                               : args[11].dat->elem_size) *
+          start[0] * args[11].stencil->stride[0];
+  base11 = base11 +
+           (long long int)(block->instance->OPS_soa ? args[11].dat->type_size
+                                                    : args[11].dat->elem_size) *
+               args[11].dat->size[0] * start[1] * args[11].stencil->stride[1];
+#ifdef OPS_GPU
   double *p_a11 = (double *)((char *)args[11].data_d + base11);
   #else
   double *p_a11 = (double *)((char *)args[11].data + base11);
