+++ apps/c/CloverLeaf_3D_HDF5/CUDA/advec_cell_kernel1_ydir_cuda_kernel.cu
@@ -141,19 +141,22 @@ void ops_par_loop_advec_cell_kernel1_ydir_execute(ops_kernel_descriptor *desc) {
   dim3 grid( (x_size-1)/block->instance->OPS_block_size_x+ 1, (y_size-1)/block->instance->OPS_block_size_y + 1, (z_size-1)/block->instance->OPS_block_size_z +1);
   dim3 tblock(block->instance->OPS_block_size_x,block->instance->OPS_block_size_y,block->instance->OPS_block_size_z);
 
-
-
-  int dat0 = (block->instance->OPS_soa ? args[0].dat->type_size : args[0].dat->elem_size);
-  int dat1 = (block->instance->OPS_soa ? args[1].dat->type_size : args[1].dat->elem_size);
-  int dat2 = (block->instance->OPS_soa ? args[2].dat->type_size : args[2].dat->elem_size);
-  int dat3 = (block->instance->OPS_soa ? args[3].dat->type_size : args[3].dat->elem_size);
-  int dat4 = (block->instance->OPS_soa ? args[4].dat->type_size : args[4].dat->elem_size);
+  long long int dat0 = (block->instance->OPS_soa ? args[0].dat->type_size
+                                                 : args[0].dat->elem_size);
+  long long int dat1 = (block->instance->OPS_soa ? args[1].dat->type_size
+                                                 : args[1].dat->elem_size);
+  long long int dat2 = (block->instance->OPS_soa ? args[2].dat->type_size
+                                                 : args[2].dat->elem_size);
+  long long int dat3 = (block->instance->OPS_soa ? args[3].dat->type_size
+                                                 : args[3].dat->elem_size);
+  long long int dat4 = (block->instance->OPS_soa ? args[4].dat->type_size
+                                                 : args[4].dat->elem_size);
 
   char *p_a[5];
 
   //set up initial pointers
-  int base0 = args[0].dat->base_offset + 
-           dat0 * 1 * (start[0] * args[0].stencil->stride[0]);
+  long long int base0 = args[0].dat->base_offset +
+                        dat0 * 1 * (start[0] * args[0].stencil->stride[0]);
   base0 = base0+ dat0 *
     args[0].dat->size[0] *
     (start[1] * args[0].stencil->stride[1]);
@@ -163,8 +166,8 @@ void ops_par_loop_advec_cell_kernel1_ydir_execute(ops_kernel_descriptor *desc) {
     (start[2] * args[0].stencil->stride[2]);
   p_a[0] = (char *)args[0].data_d + base0;
 
-  int base1 = args[1].dat->base_offset + 
-           dat1 * 1 * (start[0] * args[1].stencil->stride[0]);
+  long long int base1 = args[1].dat->base_offset +
+                        dat1 * 1 * (start[0] * args[1].stencil->stride[0]);
   base1 = base1+ dat1 *
     args[1].dat->size[0] *
     (start[1] * args[1].stencil->stride[1]);
@@ -174,8 +177,8 @@ void ops_par_loop_advec_cell_kernel1_ydir_execute(ops_kernel_descriptor *desc) {
     (start[2] * args[1].stencil->stride[2]);
   p_a[1] = (char *)args[1].data_d + base1;
 
-  int base2 = args[2].dat->base_offset + 
-           dat2 * 1 * (start[0] * args[2].stencil->stride[0]);
+  long long int base2 = args[2].dat->base_offset +
+                        dat2 * 1 * (start[0] * args[2].stencil->stride[0]);
   base2 = base2+ dat2 *
     args[2].dat->size[0] *
     (start[1] * args[2].stencil->stride[1]);
@@ -185,8 +188,8 @@ void ops_par_loop_advec_cell_kernel1_ydir_execute(ops_kernel_descriptor *desc) {
     (start[2] * args[2].stencil->stride[2]);
   p_a[2] = (char *)args[2].data_d + base2;
 
-  int base3 = args[3].dat->base_offset + 
-           dat3 * 1 * (start[0] * args[3].stencil->stride[0]);
+  long long int base3 = args[3].dat->base_offset +
+                        dat3 * 1 * (start[0] * args[3].stencil->stride[0]);
   base3 = base3+ dat3 *
     args[3].dat->size[0] *
     (start[1] * args[3].stencil->stride[1]);
@@ -196,8 +199,8 @@ void ops_par_loop_advec_cell_kernel1_ydir_execute(ops_kernel_descriptor *desc) {
     (start[2] * args[3].stencil->stride[2]);
   p_a[3] = (char *)args[3].data_d + base3;
 
-  int base4 = args[4].dat->base_offset + 
-           dat4 * 1 * (start[0] * args[4].stencil->stride[0]);
+  long long int base4 = args[4].dat->base_offset +
+                        dat4 * 1 * (start[0] * args[4].stencil->stride[0]);
   base4 = base4+ dat4 *
     args[4].dat->size[0] *
     (start[1] * args[4].stencil->stride[1]);
